#1816 fix pluploader builder file
